requirements.txt
Flask==2.3.3
flask-cors==3.0.10
google-generativeai==0.3.2
python-dotenv==0.19.0
gunicorn==20.1.0
